+++ changelog.md
@@ -1,2 +1,3 @@
 # v0.0.1 (5/26/2016-)
 * First implementation.
+* Fixed a bug where the text was not centered and a name with too many delimiters did not work.